Fixed Ignore OS Logic
The Ignore OS flag was not being set to a default causing an
"if" error because the $IGNOREOS was not being set properly.

Fixed the logic for checking for $IGNOREOS. The script was
checking for /etc/lsb-release before checking if $IGNOREOS
was set. Reordered the logic to check for $IGNOREOS first.